refactor(formatter): extract colors, table, and search formatters (#677)
